refactor(state-db): move MIGRATIONS registry verbatim to core/state/migrations (D2 step 2)
Move the 707-line MIGRATIONS array + runMigrations into core/state/migrations.ts
as a SINGLE ordered literal (no per-repo fragments — migration 001 creates 3 tables
in one un-splittable, un-renumberable fragment; splitting would corrupt
schema_migrations on deployed DBs). openStateDatabase imports runMigrations;
re-exported from state-db.ts for existing importers. Append-only contract preserved
by construction. Characterization test (exact id-list + DDL snapshot + idempotency)
green — byte-identical behavior.
